Refactor extension filters UI into dedicated widget
